fix: add label once create/update app resources
